Text.Hamlet

Plain HTML

type Html = Markup

shamlet :: QuasiQuoter

shamletFile :: FilePath -> Q Exp

xshamlet :: QuasiQuoter

xshamletFile :: FilePath -> Q Exp

Hamlet

type HtmlUrl url = Render url -> Html

A function generating an Html given a URL-rendering function.

hamlet :: QuasiQuoter

hamletFile :: FilePath -> Q Exp

xhamlet :: QuasiQuoter

xhamletFile :: FilePath -> Q Exp

I18N Hamlet

type HtmlUrlI18n msg url = Translate msg -> Render url -> Html

A function generating an Html given a message translator and a URL rendering function.

ihamlet :: QuasiQuoter

ihamletFile :: FilePath -> Q Exp

Type classes

class ToAttributes a where

Convert some value to a list of attribute pairs.

Methods

toAttributes :: a -> [(Text, Text)]

Internal, for making more

data HamletSettings

Settings for parsing of a hamlet document.

Constructors

HamletSettings

Fields

hamletDoctype :: String

The value to replace a "!!!" with. Do not include the trailing newline.

hamletNewlines :: NewlineStyle

Should we add newlines to the output, making it more human-readable? Useful for client-side debugging but may alter browser page layout.

hamletCloseStyle :: String -> CloseStyle

How a tag should be closed. Use this to switch between HTML, XHTML or even XML output.

hamletDoctypeNames :: [(String, String)]

Mapping from short names in "$doctype" statements to full doctype.

data NewlineStyle

Constructors

NoNewlines
never add newlines

NewlinesText
add newlines between consecutive text lines

AlwaysNewlines
add newlines everywhere

DefaultNewlineStyle

hamletWithSettings :: Q HamletRules -> HamletSettings -> QuasiQuoter

hamletFileWithSettings :: Q HamletRules -> HamletSettings -> FilePath -> Q Exp

defaultHamletSettings :: HamletSettings

Defaults settings: HTML5 doctype and HTML-style empty tags.

xhtmlHamletSettings :: HamletSettings

data Env

Constructors

Env

Fields

urlRender :: Maybe ((Exp -> Q Exp) -> Q Exp)

msgRender :: Maybe ((Exp -> Q Exp) -> Q Exp)

data HamletRules

Constructors

HamletRules

Fields

hrFromHtml :: Exp

hrWithEnv :: (Env -> Q Exp) -> Q Exp

hrEmbed :: Env -> Exp -> Q Exp

hamletRules :: Q HamletRules

ihamletRules :: Q HamletRules

htmlRules :: Q HamletRules

data CloseStyle

Constructors

NoClose

CloseInside

CloseSeparate