Text.Hamlet.RT

Description

Most everything exported here is exported also by Text.Hamlet. The exceptions to that rule should not be necessary for normal usage.

Public API

newtype HamletRT

Constructors

HamletRT [SimpleDoc]

data HamletData url

Constructors

HDHtml Html

HDUrl url

HDUrlParams url [(Text, Text)]

HDTemplate HamletRT

HDBool Bool

HDMaybe (Maybe (HamletMap url))

HDList [HamletMap url]

type HamletMap url = [([String], HamletData url)]

data HamletException

Constructors

HamletParseException String

HamletUnsupportedDocException Doc

HamletRenderException String

parseHamletRT :: Failure HamletException m => HamletSettings -> String -> m HamletRT

renderHamletRT :: Failure HamletException m => HamletRT -> HamletMap url -> (url -> [(Text, Text)] -> Text) -> m Html

renderHamletRT' :: Failure HamletException m => Bool -> HamletRT -> HamletMap url -> (url -> [(Text, Text)] -> Text) -> m Html

data SimpleDoc

Constructors

SDRaw String

SDVar [String]

SDUrl Bool [String]

SDTemplate [String]

SDForall [String] String [SimpleDoc]

SDMaybe [String] String [SimpleDoc] [SimpleDoc]

SDCond [([String], [SimpleDoc])] [SimpleDoc]